+++ src/main/java/net/thedevilthing/firstmod/Firstmod.java
@@ -25,7 +25,6 @@ import net.thedevilthing.firstmod.item.ModItems;
 import net.thedevilthing.firstmod.loot.ModLootConditions;
 import net.thedevilthing.firstmod.loot.ModLootModifiers;
 import net.thedevilthing.firstmod.screen.ModMenuTypes;
-import net.thedevilthing.firstmod.screen.custom.InfuserMenu;
 import net.thedevilthing.firstmod.screen.custom.InfuserScreen;
 import net.thedevilthing.firstmod.util.datapack.DataPackGenerator;
 import net.thedevilthing.firstmod.util.datapack.DataPackManager;
@@ -99,7 +98,7 @@ public class Firstmod {
     }
 
     // You can use EventBusSubscriber to automatically register all static methods in the class annotated with @SubscribeEvent
-    @EventBusSubscriber(modid = MOD_ID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+    @EventBusSubscriber(modid = MOD_ID, value = Dist.CLIENT)
     public static class ClientModEvents {
         @SubscribeEvent
         public static void onClientSetup(FMLClientSetupEvent event) {