+++ src/main/java/net/thedevilthing/firstmod/datagen/DataGenerators.java
@@ -16,7 +16,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-@EventBusSubscriber(modid = Firstmod.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
+@EventBusSubscriber(modid = Firstmod.MOD_ID)
 public class DataGenerators {
 
     @SubscribeEvent