+++ src/main/java/net/thedevilthing/firstmod/item/custom/SeedItem.java
@@ -3,7 +3,6 @@ package net.thedevilthing.firstmod.item.custom;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.component.DataComponentType;
-import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
@@ -12,13 +11,13 @@ import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.thedevilthing.firstmod.Config;
-import net.thedevilthing.firstmod.InfusionItemsConfig;
-import net.thedevilthing.firstmod.block.ModBlocks;
 import net.thedevilthing.firstmod.component.ModDataComponents;
 
 import java.util.List;
@@ -29,8 +28,6 @@ public class SeedItem extends Item{
         super(properties.stacksTo(1));
     }
 
-    public static final TagKey<Item> INFUSABLES_TAG = TagKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath("firstmod", "infusables"));
-
     private static final DataComponentType<Float> INFUSE_PERCENT = ModDataComponents.INFUSE_PERCENT.get();
     private static final DataComponentType<Item> ITEM_TYPE = ModDataComponents.ITEM_TYPE.get();
     private static final DataComponentType<Integer> TIER = ModDataComponents.TIER.get();
@@ -40,7 +37,7 @@ public class SeedItem extends Item{
         super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
 
         Float infuse_percent = stack.getOrDefault(INFUSE_PERCENT, 0.0f);
-        Item itemType = stack.getOrDefault(ITEM_TYPE, this);
+        Item itemType = stack.getOrDefault(ITEM_TYPE, Items.AIR);
         int tier = stack.getOrDefault(TIER, 1);
 
         tooltipComponents.add(Component.translatable("tooltip.firstmod.seed.tier").withStyle(ChatFormatting.GRAY)
@@ -64,80 +61,9 @@ public class SeedItem extends Item{
         ItemStack seedStack = player.getMainHandItem();
         ItemStack offHandStack = player.getOffhandItem();
 
-        float currentInfusePercent = seedStack.getOrDefault(INFUSE_PERCENT, 0.0f);
         int seedTier = seedStack.getOrDefault(TIER, 1);
 
-        // Ensure the clicked block is the specific infusion block
-        if(state.getBlock() != ModBlocks.INFUSER.get()) {
-            return InteractionResult.FAIL;
-        }
-
-        if (!offHandStack.isEmpty() && offHandStack.is(INFUSABLES_TAG)) {
-            if (currentInfusePercent >= 100.0f) {
-                player.displayClientMessage(Component.translatable("message.firstmod.infusion_complete").withStyle(ChatFormatting.GREEN), true);
-                return InteractionResult.FAIL;
-            }
-
-            Item infusedItem = offHandStack.getItem();
-            String itemID = BuiltInRegistries.ITEM.getKey(infusedItem).toString();
-
-            int itemTier = InfusionItemsConfig.getItemTier(itemID);
-
-            int availableItems = offHandStack.getCount();
-
-            if(itemTier == -1) {
-                return InteractionResult.FAIL;
-            }
-
-            if(itemTier > seedTier) {
-                player.displayClientMessage(Component.translatable("message.firstmod.high_tier",
-                                infusedItem.getDescription())
-                        .withStyle(ChatFormatting.YELLOW), true);
-
-                return InteractionResult.FAIL;
-            }
-
-            // If it's a new infusion, set the item type
-            if (currentInfusePercent == 0.0f) {
-                seedStack.set(ITEM_TYPE, infusedItem);
-            } else {
-                Item previousItem = seedStack.get(ITEM_TYPE);
-                // Ensure the infusion continues with the same item type
-                if (previousItem != infusedItem) {
-                    assert previousItem != null;
-                    player.displayClientMessage(Component.translatable("message.firstmod.wrong_item", previousItem.getDescription()).withStyle(ChatFormatting.YELLOW), true);
-                    return InteractionResult.FAIL;
-                }
-            }
-
-            int tierInfusionAmount = Config.getInfusionAmountForTier(itemTier);
-            double baseInfusePercent = Config.baseInfusion;
-            float newInfusePercent;
-            int itemsUsed = 0;
-            int tierInfusePercent = (int) Math.ceil(tierInfusionAmount * baseInfusePercent);
-
-            if (player.isShiftKeyDown()) {
-                float neededIncrease = 100.0f - currentInfusePercent;
-                itemsUsed = (int) Math.ceil(neededIncrease / tierInfusePercent);
-                if (itemsUsed > availableItems) {
-                    itemsUsed = availableItems;
-                }
-                newInfusePercent = Math.min(currentInfusePercent + (itemsUsed * tierInfusePercent), 100.0f);
-            } else {
-                newInfusePercent = Math.min(currentInfusePercent + tierInfusePercent, 100.0f);
-                itemsUsed = 1;
-            }
-
-            seedStack.set(INFUSE_PERCENT, newInfusePercent);
-
-            if (currentInfusePercent < 100.0f) {
-                offHandStack.shrink(itemsUsed);
-            } else {
-                player.displayClientMessage(Component.translatable("message.firstmod.infusion_complete").withStyle(ChatFormatting.GREEN), true);
-            }
-
-            return InteractionResult.SUCCESS;
-        } else if (!offHandStack.isEmpty() && offHandStack.getItem() instanceof SeedItem) {
+        if (!offHandStack.isEmpty() && offHandStack.getItem() instanceof SeedItem) {
             float offHandInfusePercent = offHandStack.getOrDefault(INFUSE_PERCENT, 0.0f);
 
             if (offHandInfusePercent > 0.0f && !player.isShiftKeyDown()) {
@@ -171,8 +97,14 @@ public class SeedItem extends Item{
 
             offHandStack.shrink(1);
 
-            player.displayClientMessage(Component.translatable("message.firstmod.tier_upgraded",
-                    newTier).withStyle(ChatFormatting.GREEN), true);
+            player.displayClientMessage(Component.translatable("message.firstmod.tier_upgraded", newTier).withStyle(ChatFormatting.GREEN), true);
+
+            return InteractionResult.SUCCESS;
+        } else if (offHandStack.isEmpty() && state.getBlock() == Blocks.GRINDSTONE && player.isCrouching()) {
+            seedStack.set(ITEM_TYPE, Items.AIR);
+            seedStack.set(INFUSE_PERCENT, 0.0f);
+
+            player.displayClientMessage(Component.translatable("message.firstmod.wipe_seed"), true);
 
             return InteractionResult.SUCCESS;
         }