+++ Change Log.md
@@ -135,3 +135,14 @@ Added:
 Changed:
 * Removed adding/removing of the item to Infuser's inventory on right-clicking.
 * Can add/remove item from inventory from the menu now.
+
+11/07/2025
+
+Added:
+* Infusion can be removed by shift right-clicking the grindstone.
+
+Changed:
+* Seeds are now infused in the infuser.
+
+Removed:
+* Deprecated use of bus, as now its automatically defined.